invoices+channeldb: addressing review comments.
In this commit the returned tuple of FetchAllInvoicesWithPaymentHash
is changed to only return an array os InvoiceWithPaymentHash and an
error instead of three items.

Some comments are changed and reformatted to reflect behavior better.

InvoiceExpiryWatcher.PrefetchInvoices() now returns an error, unless
there are no invoices yet stored in the DB.