channeldb: updateInvoice to use InvoiceUpdater to apply changes to DB
This commit extracts the previously separated DB updates under a DB
updater specific to kvdb and adds the `InvoiceUpdater` interface. This
way we'll be able to fully separate the steps of updating an invoice to
applying the in-memory update then storing changes to the DB that's
hidden behind the updater interface. This in turn allows us to move all
invoice update related functionality to the invoices package later on.