invoices: completely remove the update context from InvoiceUpdater
As part of a gradual refactor InvoiceUpdater will expose methods to
track gradual changes to the invoice's state. In this preparation commit
we start this by removing the context dependency from the updater and
instead replacing it with appropriate methods. Since the "legacy"
context is only important for the kv implementation we also move some
functionality over there.